requirements.txt
typer>=0.9.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
orjson>=3.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
chromadb>=0.5.5
sentence-transformers>=2.7.0
discord.py>=2.3.2
jsonschema>=4.0.0
python-dotenv>=1.0.0
alembic>=1.11.1
pytest>=7.0.0
